fixed int to double

+++ src/battle/src/moves.c
@@ -6,7 +6,7 @@
 #include <stdio.h>
 
 /** See moves.h */
-int init_class(class_t *class, enum class cl, char* info, enum stats st, int bonus)
+int init_class(class_t *class, enum class cl, char* info, enum stats st, double bonus)
 {
     assert(class != NULL);
     assert(info != NULL);
@@ -24,7 +24,7 @@ int init_class(class_t *class, enum class cl, char* info, enum stats st, int bon
 }
 
 /** See moves.h */
-class_t *new_class(enum class cl, char* info, enum stats st, int bonus)
+class_t *new_class(enum class cl, char* info, enum stats st, double bonus)
 {
     class_t *class = (class_t*) malloc(sizeof(class_t));
     int rc;
@@ -51,19 +51,19 @@ int main()
 {
     //Creates test bard class
     char* bard_des = "Charismatic, always has a joke, song, or moving speech ready";
-    class_t *test_bard = new_class(0, bard_des, 3, 2);
+    class_t *test_bard = new_class(0, bard_des, 3, 2.0);
 
     //Creates test cleric class
     char* cleric_des = "Fueled by divine inspiration, devout to the craft";
-    class_t *test_cleric = new_class(1, cleric_des, 1, 2);
+    class_t *test_cleric = new_class(1, cleric_des, 1, 2.0);
 
     //Creates test paladin class
     char* paladin_des = "Driven and committed to justice";
-    class_t *test_paladin = new_class(2, paladin_des, 2, 2);
+    class_t *test_paladin = new_class(2, paladin_des, 2, 2.0);
 
     //Creates test wizard class
     char* wizard_des = "Draws power from nature";
-    class_t *test_wizard = new_class(3, wizard_des, 0, 2);
+    class_t *test_wizard = new_class(3, wizard_des, 0, 2.0);
 
     printf("Success\n");
 }
